backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
anthropic==0.40.0
PyGithub==2.5.0
python-dotenv==1.0.1
websockets==13.1
pydantic==2.10.3
aiofiles==24.1.0
